requirements.txt
matplotlib
Django~=3.0.5
django_redis~=4.9.0
django-import-export~=2.0.2
django-extensions~=2.2.9
django-colorfield~=0.3.0
PuLP~=2.1
psycopg2-binary~=2.8.5
requests~=2.23.0
numpy
pyclustering
asgiref~=3.2.7 # comes with django
daphne~=2.4.1
channels-redis~=2.4.2
channels~=2.4.0
python-memcached~=1.59
isoweek~=1.3.3
django-ical==1.7.3
ipython
openpyxl
djangorestframework
djangorestframework-csv
django-rest-auth
django-rest-swagger
Markdown~=3.1.1
django-debug-toolbar
zeep
tdqm
django-debug-toolbar-request-history
django-filter
drf-yasg
django-cors-headers
pyprof2calltree
django-crontab
python-dateutil
